requirements.txt
openai
ollama
transformers
datasets
pandas 
tqdm 
aiohttp
jupyter
notebook
python-dotenv